% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BootPred-package.R,
%   R/bootstrap_predict-package.R
\docType{package}
\name{BootPred-package}
\alias{BootPred}
\alias{BootPred-package}
\title{BootPred: Model predictions by bootstrap}
\description{
Model predictions by bootstrap.

Model predictions by bootstrap.
}
\keyword{internal}
