% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_predict.R
\name{bootstrap_predict.glm}
\alias{bootstrap_predict.glm}
\title{Model predictions by bootstrap for glm object}
\usage{
bootstrap_predict.glm(
  object,
  newdata,
  interval = NULL,
  level = 0.95,
  bootstrap.n = 10000
)
}
\arguments{
\item{object}{a model object for which prediction is desired.}

\item{newdata}{An optional data frame in which to look for variables with which to predict. If omitted, the fitted values are used.}

\item{interval}{Type of interval calculation. "prediction" and "confidence" are supported.}

\item{level}{Tolerance/confidence level.}

\item{bootstrap.n}{number of simulation}
}
\description{
Model predictions by bootstrap for glm object
}
