% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_predict.R
\name{bootstrap_predict}
\alias{bootstrap_predict}
\title{Model predictions by bootstrap}
\usage{
bootstrap_predict(
  object,
  newdata,
  interval = NULL,
  level = 0.95,
  bootstrap.n = 10000
)
}
\arguments{
\item{object}{a model object for which prediction is desired. "lm", "glm" and "glmmTMB" are supported.}

\item{newdata}{An optional data frame in which to look for variables with which to predict. If omitted, the fitted values are used.}

\item{interval}{Type of interval calculation. "prediction" and "confidence" are supported.}

\item{level}{Tolerance/confidence level.}

\item{bootstrap.n}{number of simulation}
}
\description{
Model predictions by bootstrap
}
\examples{
x <- rnorm(15)
y <- x + rnorm(15)
new <- data.frame(x = seq(-3, 3, 0.5))
predict(lm(y ~ x), new, interval = "prediction")

}
