% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_predict.R
\name{dbetabinom}
\alias{dbetabinom}
\title{Give the density of beta-binomial distribution}
\usage{
dbetabinom(x, size, mu, phi, log = F)
}
\arguments{
\item{x}{vector of quantiles.}

\item{size}{number of trials (zero or more).}

\item{mu}{mean, defined by shape1/(shape1 + shape2) in beta distribution}

\item{phi}{dispersion patameter, defined by shape1 + shape2 in beta distribution}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}
}
\description{
Give the density of beta-binomial distribution
}
\examples{
dbetabinom(0:10, 10, 1/3, 7)

}
