% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_predict.R
\name{pbetabinom}
\alias{pbetabinom}
\title{Give the distribution function of beta-binomial distribution}
\usage{
pbetabinom(q, size, mu, phi, lower.tail = T, log.p = F)
}
\arguments{
\item{q}{vector of quantiles.}

\item{size}{number of trials (zero or more).}

\item{mu}{mean, defined by shape1/(shape1 + shape2) in beta distribution}

\item{phi}{dispersion patameter, defined by shape1 + shape2 in beta distribution}

\item{lower.tail}{logical; if TRUE (default), probabilities are P[X ≤ x], otherwise, P[X > x].}

\item{log.p}{logical; if TRUE, probabilities p are given as log(p).}
}
\description{
Give the distribution function of beta-binomial distribution
}
\examples{
pbetabinom(5, 10, 1/3, 7)

}
