% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_predict.R
\name{r_hbeta}
\alias{r_hbeta}
\title{Generate random deviates from hurdle beta distribution}
\usage{
r_hbeta(n, shape1, shape2, ziprob)
}
\arguments{
\item{n}{number of observations. If length(n) > 1, the length is taken to be the number required.}

\item{shape1}{shape parameters. Must be positive.}

\item{shape2}{shape parameters. Must be positive.}

\item{ziprob}{probability of zero(0 < ziprob < 1)}
}
\description{
Generate random deviates from hurdle beta distribution
}
\examples{
r_hbeta(10, 5, 3, 1/3)

}
