% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_predict.R
\name{r_hgamma}
\alias{r_hgamma}
\title{Generate random deviates from hurdle Gamma distribution}
\usage{
r_hgamma(n, shape, rate = 1, ziprob)
}
\arguments{
\item{n}{number of observations. If length(n) > 1, the length is taken to be the number required.}

\item{shape}{shape parameters. Must be positive.}

\item{rate}{rate parameters. Must be positive.}

\item{ziprob}{probability of zero(0 < ziprob < 1)}
}
\description{
Generate random deviates from hurdle Gamma distribution
}
\examples{
r_hgamma(10, 5, 1, 1/3)

}
