% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_predict.R
\name{r_hnbinom}
\alias{r_hnbinom}
\title{Generate random deviates from hurdle negative binomial distribution}
\usage{
r_hnbinom(n, size, prob, ziprob)
}
\arguments{
\item{n}{number of observations. If length(n) > 1, the length is taken to be the number required.}

\item{size}{target for number of successful trials, or dispersion parameter (the shape parameter of the gamma mixing distribution). Must be strictly positive, need not be integer.}

\item{prob}{probability of success in each trial. 0 < prob <= 1.}

\item{ziprob}{probability of zero(0 < ziprob < 1)}
}
\description{
Generate random deviates from hurdle negative binomial distribution
}
\examples{
r_hnbinom(10, 10, 1/3, 1/5)

}
\seealso{
[actuar] <https://cran.r-project.org/web/packages/actuar/index.html>
}
