% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_predict.R
\name{r_hpois}
\alias{r_hpois}
\title{Generate random deviates from hurdle poisson distribution}
\usage{
r_hpois(n, lambda, ziprob)
}
\arguments{
\item{n}{number of observations. If length(n) > 1, the length is taken to be the number required.}

\item{lambda}{vector of (non-negative) means.}

\item{ziprob}{probability of zero(0 < ziprob < 1)}
}
\description{
Generate random deviates from hurdle poisson distribution
}
\examples{
r_hpois(10, 5, 1/3)

}
\seealso{
[actuar] <https://cran.r-project.org/web/packages/actuar/index.html>
}
