% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_predict.R
\name{r_zipois}
\alias{r_zipois}
\title{Generate random deviates from zero-inflated poisson distribution}
\usage{
r_zipois(n, lambda, ziprob)
}
\arguments{
\item{n}{number of observations. If length(n) > 1, the length is taken to be the number required.}

\item{lambda}{vector of (non-negative) means.}

\item{ziprob}{probability of zero(0 < ziprob < 1) by Bernoulli distribution}
}
\description{
Generate random deviates from zero-inflated poisson distribution
}
\examples{
r_zipois(10, 5, 1/3)

}
