% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_predict.R
\name{rbetabinom}
\alias{rbetabinom}
\title{Generate random deviates from beta-binomial distribution}
\usage{
rbetabinom(n, size, mu, phi)
}
\arguments{
\item{n}{number of observations. If length(n) > 1, the length is taken to be the number required.}

\item{size}{number of trials (zero or more).}

\item{mu}{mean, defined by shape1/(shape1 + shape2) in beta distribution}

\item{phi}{dispersion patameter, defined by shape1 + shape2 in beta distribution}
}
\description{
Generate random deviates from beta-binomial distribution
}
\examples{
rbetabinom(10, 5, 1/3, 7)
rbetabinom(c(10,13), c(4,9), c(1/3, 1/5), c(7, 15))

}
